lucet-idl: reorganize some definitions
