+++ lucet-idl/src/config.rs
@@ -36,10 +36,7 @@ impl Config {
                 backend_config: self.backend_config,
             }),
 
-            Backend::Rust => Box::new(RustGenerator::new(
-                self.target,
-                self.backend_config,
-            )),
+            Backend::Rust => Box::new(RustGenerator::new(self.target, self.backend_config)),
         }
     }
 }
