+++ lucet-idl/src/pretty_writer.rs
@@ -1,4 +1,4 @@
-use super::errors::*;
+use super::error::IDLError;
 use std::cell::RefCell;
 use std::convert::Into;
 use std::io::prelude::*;
